Move `GradleProject` into a separate module (`plugin-testlib`).
