+++ tools/javadoc-prettifier/build.gradle
@@ -27,7 +27,7 @@ dependencies {
     implementation project(':base')
     implementation project(':plugin-base')
 
-    testImplementation project(':plugin-base').sourceSets.test.output // Test classes of `plugin-base`
+    testImplementation project(':plugin-testlib')
 }
 
 test.dependsOn(publishToMavenLocal,
