+++ tools/model-compiler/src/test/java/io/spine/tools/gradle/compiler/ProtoAnnotatorPluginShould.java
@@ -34,7 +34,7 @@ import io.spine.tools.compiler.annotation.check.MainDefinitionAnnotationCheck;
 import io.spine.tools.compiler.annotation.check.NestedTypeFieldsAnnotationCheck;
 import io.spine.tools.compiler.annotation.check.NestedTypesAnnotationCheck;
 import io.spine.tools.compiler.annotation.check.SourceCheck;
-import io.spine.tools.gradle.given.GradleProject;
+import io.spine.tools.gradle.GradleProject;
 import org.jboss.forge.roaster.Roaster;
 import org.jboss.forge.roaster.model.impl.AbstractJavaSource;
 import org.jboss.forge.roaster.model.source.JavaClassSource;
