+++ tools/model-compiler/src/test/java/io/spine/tools/gradle/compiler/ValidatingBuilderGenPluginShould.java
@@ -20,7 +20,7 @@
 
 package io.spine.tools.gradle.compiler;
 
-import io.spine.tools.gradle.given.GradleProject;
+import io.spine.tools.gradle.GradleProject;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
