+++ tools/model-compiler/src/test/java/io/spine/tools/gradle/compiler/ValidationRulesLookupPluginShould.java
@@ -21,7 +21,7 @@
 package io.spine.tools.gradle.compiler;
 
 import io.spine.code.DefaultProject;
-import io.spine.tools.gradle.given.GradleProject;
+import io.spine.tools.gradle.GradleProject;
 import io.spine.code.properties.PropertyFile;
 import io.spine.validate.rules.ValidationRules;
 import org.junit.Rule;
