+++ tools/plugin-testlib/src/test/resources/build.gradle
