+++ tools/plugin-testlib/src/test/resources/gradle_project_test/src/main/java/Foo.java
